+++ src/server/game/Mails/Mail.cpp
@@ -53,7 +53,7 @@ MailSender::MailSender(Object* sender, MailStationery stationery) : m_stationery
         default:
             m_messageType = MAIL_NORMAL;
             m_senderId = 0;                                 // will show mail from not existed player
-            SF_LOG_ERROR("misc", "MailSender::MailSender - Mail have unexpected sender typeid (%u)", sender->GetTypeId());
+            SF_LOG_ERROR("misc", "MailSender::MailSender - Mail have unexpected sender typeid (%u)", uint8(sender->GetTypeId()));
             break;
     }
 }
